requirements.txt
PySide6==6.11.0
PyGithub==2.1.1
GitPython==3.1.40
requests==2.31.0
keyring==24.2.0
pyperclip==1.8.2
